requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
crispy-bootstrap5==0.7
debugpy==1.6.7
decorator==5.1.1
Django==3.2.19
django-mathfilters==1.0.0
django-monthfield==0.1.4
executing==1.2.0
google-api-core==2.11.0
google-api-python-client==2.20.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.16
wcwidth==0.2.6
zcache==0.2
